common js compatible build
