+++ .gitignore
@@ -10,7 +10,7 @@ node_modules
 /coverage
 
 # production
-/build
+#/build
 
 # env 
 .env
