+++ package.json
@@ -1,9 +1,12 @@
 {
   "name": "@foursquare/foursquare-places",
-  "version": "1.0.2",
+  "version": "1.0.3",
   "description": "This is a javascript wrapper for Foursquare's APIs",
-  "main": "index.js",
+  "main": "./build/index.js",
   "scripts": {
+    "build": "NODE_ENV=production babel lib --out-dir build",
+    "clean": "rm ./build/*.js",
+    "dist": "npm run clean && npm run build",
     "test": "jest"
   },
   "repository": {
@@ -30,6 +33,7 @@
   },
   "homepage": "https://github.com/foursquare/foursquare-places#readme",
   "devDependencies": {
+    "@babel/cli": "^7.6.0",
     "@babel/core": "^7.6.0",
     "@babel/preset-env": "^7.6.0",
     "babel-jest": "^24.9.0",
